minunit and TDD

+++ days_c/03/ex01/ft_strncmp.c
@@ -13,10 +13,10 @@
 
 int		ft_strncmp(char *s1, char *s2, unsigned int n)
 {
-	int		i;
+	unsigned int	i;
 
 	i = -1;
-	while (++i < n)
+	while ((unsigned int)++i < n)
 		if (s1[i] != s2[i])
 			return (s1[i] - s2[i]);
 		else if (s1[i] == '\0')
